requirements.txt
anyio==4.9.0
apsw==3.49.1.0
apswutils==0.0.2
asttokens==3.0.0
babel==2.17.0
beautifulsoup4==4.13.3
bibtexparser==1.4.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
executing==2.2.0
fastcore==1.8.0
fastlite==0.1.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
icecream==2.1.4
idna==3.10
itsdangerous==2.2.0
lxml==5.3.1
marko==2.1.2
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psutil==7.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-fasthtml==0.12.5
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vl-convert-python==1.7.0
watchfiles==1.0.4
webencodings==0.5.1
websockets==15.0.1
